requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==1.57.0
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
faiss-cpu==1.9.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
tiktoken==0.8.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
